다른 마이크로 서비스에도 카프카용 설정 클래스 추가.
