+++ read-diary-service/src/main/java/com/dasd412/api/readdiaryservice/config/KafkaConfiguration.java
@@ -1,0 +1,14 @@
+package com.dasd412.api.readdiaryservice.config;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Configuration;
+
+@ConditionalOnProperty(value = "kafka.enabled", matchIfMissing = true)
+@Configuration
+public class KafkaConfiguration {
+
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
+}
