+++ writer-service/src/main/resources/application-test.properties
@@ -1,0 +1,1 @@
+kafka.enabled=false
